requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==4.0.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blinker==1.8.2
cachetools==5.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.1
fastjsonschema==2.16.3
filelock==3.16.1
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==2.0.2
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.25.1
idna==3.4
image-classifiers==1.0.0
importlib_metadata==8.5.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
Keras-Applications==1.0.8
kiwisolver==1.4.4
langchain==0.0.184
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.22.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
multidict==6.1.0
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.9.1
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.10.1
numpy==1.26.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.1
plotly==5.24.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
pydantic==1.10.18
pydeck==0.9.1
Pygments==2.14.0
Pympler==1.1
pyodbc==4.0.39
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
rake-nltk==1.0.5
regex==2024.9.11
reportlab==4.0.4
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.16.0
semver==3.0.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.17
stack-data==0.6.2
starlette==0.26.1
streamlit==1.18.1
tenacity==8.5.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.2
tqdm==4.66.5
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uri-template==1.2.0
urllib3==2.0.3
uvicorn==0.22.0
validators==0.34.0
vit-keras==0.1.2
watchdog==5.0.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.6
wsproto==1.2.0
XlsxWriter==3.1.2
yarl==1.12.1
zipp==3.20.2
streamlit==1.18.1
streamlit_lottie==0.0.5
